fix: 将类别ID的数据类型从Integer改为Long
修改了TodoMapper.xml、StatisticsService接口及其实现类、TodoMapper接口和ClientStatisticsController中与类别ID相关的数据类型，从Integer统一改为Long，以支持更大的ID范围

+++ src/main/java/com/backend/controller/client/ClientStatisticsController.java
@@ -31,7 +31,7 @@ public class ClientStatisticsController {
     public ResultBean<List<Map<String, Object>>> getTodoCountByCategory(@RequestBody Map<String, Object> params) {
         String startDate = (String) params.get("startDate");
         String endDate = (String) params.get("endDate");
-        List<Integer> categoryIdList = (List<Integer>) params.get("categoryIdList");
+        List<Long> categoryIdList = (List<Long>) params.get("categoryIdList");
         return ResultBean.success(statisticsService.getTodoCountByCategory(startDate, endDate, categoryIdList));
     }
     
@@ -44,7 +44,7 @@ public class ClientStatisticsController {
     public ResultBean<List<Map<String, Object>>> getTodoCountByCategoryAndDate(@RequestBody Map<String, Object> params) {
         String startDate = (String) params.get("startDate");
         String endDate = (String) params.get("endDate");
-        List<Integer> categoryIdList = (List<Integer>) params.get("categoryIdList");
+        List<Long> categoryIdList = (List<Long>) params.get("categoryIdList");
         return ResultBean.success(statisticsService.getTodoCountByCategoryAndDate(startDate, endDate, categoryIdList));
     }
 }
